It finally works on my machine.
For a reason I don't fully understand, the global version of babel on
my machine is ^6 but the component keeps looking for version 5 and is
trying to use version 5 with the version 6 babel-core, babel-loader,
etc.  This breaks everything.  By explicitly setting the version of
babel in the example/package.json to ^6 it works.  It was in the print
out of the 3rd error that I saw the mismatch.  Lucas helped a lot with
debugging.